Move ResponseType interface to `responsetypes` package
Move ResponseType interface to `responsetypes` package

[#463]
