+++ core/src/main/java/org/axonframework/queryhandling/responsetypes/ListResponseType.java
@@ -1,7 +1,5 @@
 package org.axonframework.queryhandling.responsetypes;
 
-import org.axonframework.queryhandling.ResponseType;
-
 import java.lang.reflect.Type;
 import java.util.List;
