+++ core/src/main/java/org/axonframework/queryhandling/responsetypes/PageResponseType.java
@@ -1,7 +1,5 @@
 package org.axonframework.queryhandling.responsetypes;
 
-import org.axonframework.queryhandling.ResponseType;
-
 import java.lang.reflect.Type;
 
 /**
