+++ core/src/main/java/org/axonframework/queryhandling/responsetypes/ResponseType.java
@@ -1,4 +1,4 @@
-package org.axonframework.queryhandling;
+package org.axonframework.queryhandling.responsetypes;
 
 import java.lang.reflect.Type;
